+++ src/services/user_service.rs
@@ -130,8 +130,8 @@ impl UserService {
         user_id: i64,
         params: &PaginationParams,
     ) -> AppResult<PaginatedResponse<UserResponse>> {
-        let offset = params.get_offset() as i64;
-        let limit = params.get_limit() as i64;
+        let offset = params.get_offset();
+        let limit = params.get_limit();
 
         // 获取总数
         let total: i64 = sqlx::query_scalar("SELECT COUNT(*) FROM users WHERE referrer_id = $1")
